backend refactor with optimistation done on data collection pipelines
