--- backend/__init__.py
@@ -1,25 +1,0 @@
-# Import dependencies
-from .functions import (
-    TrackManAggregator,
-    configure_logging,
-    ScorecardParser,
-    RoundAggregator,
-    SeleniumDriver,
-    TrackMan
-)
-from .interfaces import (
-    AbstractDataCollection,
-    AbstractSeleniumDriver
-)
-
-__all__ = [
-    "AbstractDataCollection",
-    "AbstractSeleniumDriver",
-    "TrackManAggregator",
-    "configure_logging",
-    "ScorecardParser",
-    "RoundAggregator",
-    "SeleniumDriver",
-    "RoundData",
-    "TrackMan"
-]
