--- backend/functions/__init__.py
@@ -1,18 +1,0 @@
-# Import project dependencies
-from .trackman import TrackManAggregator, TrackMan
-from .scorecard_aggregator import RoundAggregator
-from .scorecard_parser import ScorecardParser
-from .selenium_driver import SeleniumDriver
-from .scorecard import Hole19Navigator
-from .logging import configure_logging
-
-
-__all__ = [
-    "TrackManAggregator",
-    "configure_logging",
-    "ScorecardParser",
-    "RoundAggregator",
-    "Hole19Navigator",
-    "SeleniumDriver",
-    "TrackMan"
-]
