+++ backend/functions/scorecard_parser.py
@@ -1,12 +1,12 @@
 # Import dependencies
 from selenium.webdriver.common.by import By
 from .selenium_driver import SeleniumDriver
+from shared import Variables, BlobClient
 from datetime import datetime, date
-from shared import Variables
 import logging
 import re
 
-class ScorecardParser(SeleniumDriver):
+class ScorecardParser(SeleniumDriver, BlobClient):
     """
     Parses golf scorecards from the Hole19 website.
 
@@ -377,6 +377,18 @@ class ScorecardParser(SeleniumDriver):
 
         return results
 
+    def identify_new_data(self, scorecard_urls: list) -> list:
+        """
+        """
+        scorecard_ids = [url.split("/")[-1] for url in scorecard_urls]
+
+        collected_scorecard = self.list_blob_filenames(container_name="golf", directory_path="scorecards")
+
+        collected_scorecard_ids = [file.split("_")[-1].replace(".json", "") for file in collected_scorecard]
+
+        return [f"https://www.hole19golf.com/performance/rounds/{id}"
+                for id in list(set(scorecard_ids) - set(collected_scorecard_ids))]
+
     def collect_scorecard_data(self, url: str) -> tuple[list[dict], str]:
         """
         Collect and process scorecard data from a URL.
@@ -395,7 +407,7 @@ class ScorecardParser(SeleniumDriver):
         self.driver.get(url)
         round_date = self.get_round_date()
         course_name = self.get_course_name()
-        file_name = f'scorecards/{course_name}_{round_date}.json'
+        file_name = f'scorecards/{course_name}_{round_date}_{url.split("/")[-1]}.json'
 
         # Collect scorecard data
         scorecard_section = self.driver.find_element(By.CSS_SELECTOR, 'section.round-scorecard')
